+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # cpkg
 
+## 0.2.2.0
+
+  * Add `printLdLibFlags` function and add functionality to CLI interface
+  * `cfg.installDir` is now absolute
+
+## 0.2.1.0
+
+  * Add `SymlinkManpage` command
+
 ## 0.2.0.1
 
   * Pass `-p0` option to `patch`